requirements/prod.txt
bcrypt==3.2.0
fastapi==0.63.0
passlib==1.7.4
psycopg2-binary==2.8.6
pydantic==1.7.3
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
SQLAlchemy==1.3.22
starlette==0.13.6
uvicorn==0.13.3
